ALSA update   - cmipci driver cleanups (ac3 & surround)   - replaced snd_dma_residue() with snd_dma_pointer()   - GCC 3.3 warnings removal   - timer interface     - recoded using tasklet     - improved slave timer locking (should be much faster)     - added async support   - improved ioctl32 wrapper functions   - fixed Makefile problems (synth modules were build for not selected driver)   - AC97 codec     - improved SPSA control     - moved reset function outside the main init code     - improved ALC650 initialization   - USB driver     - added quirk for Roland XV-2020

+++ sound/core/info.c
@@ -796,20 +796,20 @@ int snd_info_get_line(snd_info_buffer_t * buffer, char *line, int len)
 	while (--len > 0) {
 		c = *buffer->curr++;
 		if (c == '\n') {
-			if ((buffer->curr - buffer->buffer) >= buffer->size) {
+			if ((buffer->curr - buffer->buffer) >= (long)buffer->size) {
 				buffer->stop = 1;
 			}
 			break;
 		}
 		*line++ = c;
-		if ((buffer->curr - buffer->buffer) >= buffer->size) {
+		if ((buffer->curr - buffer->buffer) >= (long)buffer->size) {
 			buffer->stop = 1;
 			break;
 		}
 	}
 	while (c != '\n' && !buffer->stop) {
 		c = *buffer->curr++;
-		if ((buffer->curr - buffer->buffer) >= buffer->size) {
+		if ((buffer->curr - buffer->buffer) >= (long)buffer->size) {
 			buffer->stop = 1;
 		}
 	}
